linux-user: Allow 64-bit offsets in mmap2() syscall on 32-bit targets
The mmap2() syscall allows 32-bit guests to specify the offset into a
mapping in 4096-byte units (instead of bytes, as is done by mmap(2)).
This enables applications that use a 32-bit off_t to map large files
(up to 2^44 bytes).

So when emulating the mmap2 syscall the offset parameter is shifted by
12 bits, which then may exceed what the abi_ulong (32-bit) type of the
offset parameter in target_mmap() can hold.

Overcome this possible issue by switching hard to uint64_t in
target_mmap() and mmap_frag().

Note: compile-tested only.

Message-Id: <20221222052021.16734-1-deller@gmx.de>

+++ linux-user/syscall.c
@@ -10097,7 +10097,7 @@ static abi_long do_syscall1(CPUArchState *cpu_env, int num, abi_long arg1,
 #endif
         ret = target_mmap(arg1, arg2, arg3,
                           target_to_host_bitmask(arg4, mmap_flags_tbl),
-                          arg5, arg6 << MMAP_SHIFT);
+                          arg5, ((uint64_t)arg6) << MMAP_SHIFT);
         return get_errno(ret);
 #endif
     case TARGET_NR_munmap:
